# Optimize `take` kernel for `BinaryViewArray` and `StringViewArray`