@a10y 2024-08-01
Looks like benchmarks are the same or better using new_unchecked

```
group                                            2905ce6796cad396                       new_unchecked
-----                                            ----------------                       -------------
take stringview 1024                             18.71    14.1±0.17µs        ? ?/sec    1.00   751.3±12.43ns        ? ?/sec
take stringview 512                              15.24     6.3±0.10µs        ? ?/sec    1.00   412.7±18.34ns        ? ?/sec
take stringview null indices 1024                15.79    11.7±0.16µs        ? ?/sec    1.00   740.6±30.60ns        ? ?/sec
take stringview null indices 512                 12.23     5.4±0.07µs        ? ?/sec    1.00   440.4±30.70ns        ? ?/sec
take stringview null values 1024                 5.45      9.1±0.08µs        ? ?/sec    1.00  1674.5±71.90ns        ? ?/sec
take stringview null values null indices 1024    2.71      6.9±0.08µs        ? ?/sec    1.00      2.6±0.42µs        ? ?/sec
```