@XiangpengHao 2024-08-01
Looks good to me! Great to know that `take` can be even faster!