@alamb 2024-08-01
It is not often we see a 16x performance improvement 🚀  -- thank you @a10y 

BTW I got some pretty sweet looking numbers for ClickBench with StringView in DataFusion (details on https://github.com/apache/datafusion/pull/11723). @XiangpengHao  mentioed that this implementation will likely make it even better. I'll see if I can run some numbers